# docs/apis/openapi.yaml
openapi: 3.0.3
info:
  title: hmos-score-agent API
  version: 0.1.0
  description: HarmonyOS 代码评分服务对外 HTTP 接口和远端 callback 契约。
servers:
  - url: http://localhost:3000
    description: 本地开发服务，默认 PORT=3000。
...
paths:
...
  /score/remote-tasks/{taskId}/human-review:
    post:
      tags:
        - Score
      summary: 提交人工复核和整单人工评级
      description: 提交逐条复核结论和整单人工评级，并根据复核结果重算分数。重复提交会覆盖 result.json 中的最新 human_review_revision。
      operationId: submitHumanReview
      parameters:
        - $ref: "#/components/parameters/TaskIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HumanReviewRequest"
            example:
              reviewer: "reviewer-001"
              manualLevel: "L2"
              overallComment: "自动评分漏掉了全屏态返回路径的问题。"
              itemReviews:
                - itemId: 1
                  agree: false
                  reason: "该复核项判断过轻，问题会影响主流程。"
              riskReviews:
                - riskId: 1
                  agree: false
                  correctedLevel: "medium"
                  reason: "当前风险应降为中风险。"
      responses:
        "200":
          description: 人工复核已接收，数据集样本已写入；存在 score_effect 时会同步重算分数。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HumanReviewResponse"
              example:
                success: true
                taskId: 4
                status: completed
                summary:
                  itemReviewCount: 1
                  riskReviewCount: 1
                  riskAgreementCount: 0
                  riskDisagreementCount: 1
                  datasetItemCount: 2
                  hasOverallComment: true
                  manualLevel: "L2"
                  autoScore: 88
                  autoRating: "L3"
                  gapQualified: true
                  analysisStatus: "completed"
                  scoreRecalculationApplied: true
                  originalTotalScore: 88
                  revisedTotalScore: 80
                  changedItemScoreCount: 0
                  changedDimensionScoreCount: 1
                message: "人工复核结果已接收，结果分数已重新计算。人工评级已接收，评分差异较大，已完成差异原因分析。"
        "400":
          description: 人工复核请求体非法，或 itemId/riskId 无法匹配当前 result.json。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseWithOptionalTaskId"
              examples:
                invalidManualLevel:
                  value:
                    success: false
                    message: "manualLevel must be one of L1, L2, L3, L4, L5, L6"
                invalidRiskId:
                  value:
                    success: false
                    taskId: 4
                    message: "riskReviews[0].riskId does not match result risks"
        "404":
          description: 任务记录或结果文件不存在。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseWithOptionalTaskId"
              example:
                success: false
                taskId: 4
                message: "Remote task not found"
        "409":
          description: 任务未完成，或 result.json 缺少重算/评级所需字段。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseWithOptionalTaskIdAndStatus"
              examples:
                notCompleted:
                  value:
                    success: false
                    taskId: 4
                    status: running
                    message: "Remote task is not completed yet"
                missingScore:
                  value:
                    success: false
                    taskId: 4
                    message: "overall_conclusion.total_score is required for manual rating gap analysis"
components:
  parameters:
    TaskIdPath:
      name: taskId
      in: path
      required: true
      description: 远端评分任务 ID。
      schema:
        type: number
  schemas:
...
    RemoteTaskRecordStatus:
      type: string
      enum:
        - preparing
        - queued
        - running
        - completed
        - failed
        - timed_out
      description: 服务记录的远端任务状态。
...
    HumanReviewRequest:
      type: object
      required:
        - manualLevel
      properties:
        reviewer:
          type: string
          description: 可选复核人标识。
        manualLevel:
          $ref: "#/components/schemas/HumanManualLevel"
        overallComment:
          type: string
          description: 可选整体评价；人工评级差异分析会将它作为评级依据。
        itemReviews:
          type: array
          description: 对 result.json.human_review_items[] 的逐条复核。缺失或空数组合法。
          items:
            $ref: "#/components/schemas/HumanReviewItemReview"
        riskReviews:
          type: array
          description: 对 result.json.risks[] 的风险等级复核。缺失或空数组合法。
          items:
            $ref: "#/components/schemas/HumanRiskReview"
    HumanReviewItemReview:
      type: object
      required:
        - itemId
        - agree
      properties:
        itemId:
          type: number
          description: result.json.human_review_items[] 中的复核项 ID；必须为正整数且不可重复。
        agree:
          type: boolean
          description: 是否同意系统当前判断。
        reason:
          type: string
          description: agree=false 时必填的原因。
    HumanRiskReview:
      type: object
      required:
        - riskId
        - agree
      properties:
        riskId:
          type: number
          description: result.json.risks[] 中的风险项 ID；必须为正整数且不可重复。
        agree:
          type: boolean
          description: 是否同意系统当前风险等级。
        correctedLevel:
          $ref: "#/components/schemas/HumanRiskLevel"
        reason:
          type: string
          description: agree=false 时必填的原因。
    HumanManualLevel:
      type: string
      enum:
        - L1
        - L2
        - L3
        - L4
        - L5
        - L6
      description: 整单人工评级。
    HumanRiskLevel:
      type: string
      enum:
        - high
        - medium
        - low
        - none
      description: 人工修正后的风险等级。
    HumanReviewResponse:
      type: object
      required:
        - success
        - taskId
        - status
        - summary
        - message
      properties:
        success:
          type: boolean
          description: 请求是否处理成功。
        taskId:
          type: number
          description: 远端任务 ID。
        status:
          type: string
          enum:
            - completed
          description: 人工复核处理状态，当前为同步完成。
        summary:
          $ref: "#/components/schemas/HumanReviewSummary"
        message:
          type: string
          description: 处理结果说明。
    HumanReviewSummary:
      type: object
      required:
        - itemReviewCount
        - riskReviewCount
        - riskAgreementCount
        - riskDisagreementCount
        - datasetItemCount
        - hasOverallComment
        - manualLevel
        - autoScore
        - autoRating
        - gapQualified
        - analysisStatus
      properties:
        itemReviewCount:
          type: number
          description: 本次提交的评分项复核数量。
        riskReviewCount:
          type: number
          description: 本次提交的风险项复核数量。
        riskAgreementCount:
          type: number
          description: 同意系统风险等级的数量。
        riskDisagreementCount:
          type: number
          description: 不同意系统风险等级的数量。
        datasetItemCount:
          type: number
          description: 写入人工复核数据集的样本数量。
        hasOverallComment:
          type: boolean
          description: 是否提供非空整体评价。
        manualLevel:
          $ref: "#/components/schemas/HumanManualLevel"
        autoScore:
          type: number
          description: result.json 中的自动评分总分。
        autoRating:
          $ref: "#/components/schemas/HumanManualLevel"
        gapQualified:
          type: boolean
          description: 是否达到人工评级差异分析阈值。
        analysisStatus:
          type: string
          enum:
            - completed
            - skipped
          description: 人工评级差异分析状态；human-review 接口达到差异分析阈值时仍返回 completed，message 会说明差异分析已转入后台执行。
        scoreRecalculationApplied:
          type: boolean
          description: 存在分数变化时返回，表示已应用重算。
        originalTotalScore:
          type: number
          description: 重算前总分；发生分数变化时返回。
        revisedTotalScore:
          type: number
          description: 重算后总分；发生分数变化时返回。
        changedItemScoreCount:
          type: number
          description: 分数变化的评分项数量；发生分数变化时返回。
        changedDimensionScoreCount:
          type: number
          description: 分数变化的维度数量；发生分数变化时返回。
    ErrorResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          description: 固定为 false。
        message:
          type: string
          description: 错误说明。
    ErrorResponseWithOptionalTaskId:
      allOf:
        - $ref: "#/components/schemas/ErrorResponse"
        - type: object
          properties:
            taskId:
              type: number
              description: 如果路径参数可解析则返回。
    ErrorResponseWithOptionalTaskIdAndStatus:
      allOf:
        - $ref: "#/components/schemas/ErrorResponseWithOptionalTaskId"
        - type: object
          properties:
            status:
              $ref: "#/components/schemas/RemoteTaskRecordStatus"